% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{fmt_currency}
\alias{fmt_currency}
\title{Currency format function factory}
\usage{
fmt_currency(prefix = "$", suffix = "", digits = 2L, big_mark = ",")
}
\arguments{
\item{prefix}{Character prepended before the number. Default \code{"$"}.}

\item{suffix}{Character appended after the number. Default \code{""}.}

\item{digits}{Integer decimal places. Default \code{2L}.}

\item{big_mark}{Thousands separator. Default \code{","}.}
}
\value{
A function \verb{function(x) -> character}.
}
\description{
Returns a formatting function for use as \code{format_fn} in \code{\link[=col_def]{col_def()}}.
}
