% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{fmt_percent}
\alias{fmt_percent}
\title{Percentage format function factory}
\usage{
fmt_percent(digits = 2L, plus = TRUE)
}
\arguments{
\item{digits}{Integer decimal places. Default \code{2L}.}

\item{plus}{Logical. Prefix non-negative values with \code{"+"}. Default \code{TRUE}.}
}
\value{
A function \verb{function(x) -> character}.
}
\description{
Returns a formatting function for use as \code{format_fn} in \code{\link[=col_def]{col_def()}}.
}
