% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{rows_to_nodes}
\alias{rows_to_nodes}
\title{Convert data frame rows into leaf nodes}
\usage{
rows_to_nodes(df, name_col, value_cols)
}
\arguments{
\item{df}{A data frame.}

\item{name_col}{Column name to use as the node label.}

\item{value_cols}{Character vector of column names to carry as \code{.values}.}
}
\value{
A list of \code{\link[=node]{node()}} objects.
}
\description{
Convert data frame rows into leaf nodes
}
